# use voxel manip for lv light illumination