As promised in #291

Timings:
* Old code: 10 to 20 ms
* New code: 2 to 8 ms

Test failures:
```
Failure → spec/lv_network_spec.lua @ 118
LV machine network charges battery box
spec/lv_network_spec.lua:121: Expected (number) 0 to be more than (number) 1000

Error → spec/lv_network_spec.lua @ 95
LV machine network executes network
machines/LV/lamp.lua:42: attempt to index global 'VoxelArea' (a nil value)

Error → spec/lv_network_spec.lua @ 104
LV machine network kills network when switching station disappear
machines/LV/lamp.lua:42: attempt to index global 'VoxelArea' (a nil value)

Error → spec/lv_network_spec.lua @ 124
LV machine network smelts ores
machines/LV/lamp.lua:42: attempt to index global 'VoxelArea' (a nil value)

Error → spec/lv_network_spec.lua @ 134
LV machine network grinds ores
machines/LV/lamp.lua:42: attempt to index global 'VoxelArea' (a nil value)

Error → spec/lv_network_spec.lua @ 144
LV machine network comperess sand
machines/LV/lamp.lua:42: attempt to index global 'VoxelArea' (a nil value)

Error → spec/lv_network_spec.lua @ 154
LV machine network cuts power when generators disappear
machines/LV/lamp.lua:42: attempt to index global 'VoxelArea' (a nil value)
```

Should be mostly mineunit related (missing `VoxelArea` var)

@S-S-X suggestions? i haven't looked into it but how much work would it be to implement the `VoxelArea` function in  mineunit?