@S-S-X 2022-11-17
> how much work would it be to implement the VoxelArea function in mineunit?

I think it might already work but only if using actual complete engine core libraries (with `mineunit --fetch-core 5.6.1 --engine-version 5.6.1`)
Never tested but discussed here: https://github.com/S-S-X/mineunit/pull/84 (could also just merge that PR to include VoxelArea in base mineunit without downloading engine libraries)