@OgelGames 2022-11-18
FYI, `VoxelArea:iterp(minp, maxp)` exists: https://github.com/minetest/minetest/blob/587f6656a4b86346e35da1b43b48b832d3f1b32e/doc/lua_api.txt#L4455